tables done, need to work on seed files
