+++ app/models/__init__.py
@@ -3,3 +3,4 @@ from .user import User
 from .db import environment, SCHEMA
 
 
+
